i3: add notification program (dunst)
The network manager was throwing some random errors due
to not having a notification daemon available in the system.
Due to this, the daemon was really slow in the start up.

By installing the dunst daemon now we have a notification
daemon for the system with the default configuration
provided by the program.